demo-api/pyproject.toml
[tool.poetry]
name = "demo-api"
version = "0.1.0"
description = ""
authors = ["Josh Hoiland <josh.hoiland@cavallo.com>"]
readme = "README.md"
packages = [{include = "demo_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
typer = "^0.12.3"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-otlp = "^1.25.0"
theine = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
